requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2
django-idempotency-key==1.2.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pytz==2023.3
result==0.9.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
